cv-tailor/requirements.txt
flask
anthropic
reportlab
pdfplumber
python-dotenv
